Move preview sprite text's position.
